requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
openai
requests
python-dotenv
langchain-text-splitters
rank_bm25
rank_bm25
PyYAML
